Version Constraints
Add support for composer.json containing version constraints. This sets the constraint for affected packages to * in local state, alongside the minimum-stability which is already set to dev. Original constraints and minimum-stability are now backed up in within composer.json and restored when going back to remote state. Update the README to reflect this change.